RELEASE: 0.6-beta6
git-svn-id: https://plugins.svn.wordpress.org/job-manager/trunk@211794 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ admin-settings.php
@@ -587,6 +587,9 @@ function jobman_icons_updatedb() {
 
 	$deletes = explode( ',', $_REQUEST['jobman-delete-list'] );
 	foreach( $deletes as $delete ) {
+		if( empty( $delete ) )
+			continue;
+		
 		wp_delete_attachment( $delete );
 		
 		unset( $options['icons'][array_search( $delete, $options['icons'] )] );
